Remove rails-specific rendering stuff
